backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
pyotp
python-multipart
pydantic